ica: Cache libica key object as data attached to key object
Cache the RSA or EC libica key object by attaching it to the key object.
Subsequent crypto operations using the same key object will reuse the
libica key object and save some time, because it is not needed to recreate
the libica key object for every request.

Care must be taken that the software fallback uses OpenSSL, and the code
in mech_openssl.c may also cache the OpenSSL PKEY object by attaching it
to the key object. Thus, struct openssl_ex_data must be the first field
of the ex_data structure used by the ICA token, and ica_free_ex_data()
must call openssl_free_ex_data() as very last.

Because attached key objects live longer, and may be concurrently used by
multiple threads, check for RSA keys with p < q and let libica correct it
if needed at time the libica key object is attached. This avoids that a
call to ica_rsa_crt() later on may correct it in a non-thread-save way.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>