configure: Don't add libcap to LIBS globally
Library libcap is only needed by pkcsslotd, but not by all the other
executables.

AC_CHECK_LIB without a 'action-if-found' specified will add the library
to the global LIBS, and thus all executables will link against it. Avoid this
by specifying a 'action-if-found' action and build a LIBCAP_LIBS variable
(which however is unused).

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>