EP11: Re-enciphering session-bound key blobs require the session to be logged in
When re-enciphering session-bound key blobs, the EP11 session must be logged
in into the APQN where the re-encipher command is performed with. In case the
session is not logged in CKR_SESSION_CLOSED is returned.

Pass the corresponding session to the re-encipher function and add a retry
logic in case of CKR_SESSION_CLOSED to re-login the session.

Re-enciphering token objects happens in a thread inside the pkcshsm_mk_change
tool, so pass the R/W USER session that the tool must have open and logged in
to the re-encipher function when iterating over the token objects.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>