EP11: Public objects can't be bound to an EP11 session
To work with secure keys bound to an EP11 session, the VHSM-session or the
EP11 session associated with the PKCS#11 session (strict session mode) must
first be logged in. This happens only when a USER session is opened, or
when a C_Login is performed for USER, but not for public sessions. This is
because the VHSM pin and the session's pin blobs are stored in private token
HW-objects and thus the USER pin is required to access the objects (which
are encrypted by the token master key).

Secure keys stored in public objects can be used on public sessions, no
login is required for that. Furthermore, they could be used in SO sessions.
SO does not have access to private objects (CK_PRIVATE=TRUE). Thus, the
related EP11 session may not be logged in when such a session bound public
key object would be used. Public objects can therefore not be bound to an
EP11 session, because they would not be usable in public or SO sessions.
Also it is not possible to login the EP11 session on a public or SO session,
since the VHSM-pin can not be obtained, and the session pin blob could not be
stored.

When creating public key objects (CKA_PRIVATE=FALSE), don't bind the secure
key to the EP11 session. Only for private keys (CKA_PRIVATE=TRUE) the secure
key can be bound to an EP11 session.

For asymmetric key pairs, the public key part and the private key part can
have different attributes, especially in regards of CKA_PRIVATE and CKA_TOKEN.
Thus, the key parts may need to be bound to different EP11 sessions, or one
key part is bound to an EP11 session while the other key part is not bound to
an EP11 session at all. This is dependent on the CKA_PRIVATE and CKA_TOKEN
attributes as well as if VHSM-mode and/or strict session mode is active.

The EP11 m_GenerateKeyPair() function binds both, the private key blob as
well as the public key (i.e. the MACed SPKI) to the same session (or both to no
session). To allow different sessions for public key part and private key part,
first the m_GenerateKeyPair() is performed with the private key part session
pin blob (or NULL if no session is desired). In case the public key part is
to be bound to a different session, then the MAC-ed SPKI is re-created by means
of an m_UnwrapKey() with CKM_IBM_TRANSPORTKEY using the public key part's
session pin blob (or NULL if no session is desired).

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>