CCA/EP11: report APQN consistency state change only once
When an APQN event causes the APQN consistency state to change to inconsistent,
report that on syslog only the first time, but not if on subsequent events
the state also stays at inconsistent.

Similar, only report the state change from inconsistent to consistent only
once, and not for every event that detects a consistent state.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>