CCA: Handle concurrent MK change for AES-XTS keys special cases
When the new CCA AES master key is activated after the first part of an
AES-XTS key has been generated or imported, but before the second part of
the AES-XTS key is generated or imported, special handling is required.

Also, when the first part of an AES-XTS key has been generated or imported
while the new CCA AES master key was already activated, but before generating
or importing the second part of the AES-XTS key the APQN with the new CCA AES
master key gets offline, and another APQN which still has the old CCA AES
master key is used to generate or import the second part of the AES-XTS key,
both key parts needs to be recreated (now both with the old CCA AES master
key).

When importing an AES-XTS from an opaque key blob, then both parts must use
the same master key (i.e. the keys contain the same master key verification
pattern). If that is not the case, then the import fails with
CKR_ATTRIBUTE_VALUE_INVALID. This check was already present in the code.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>