EP11: Ensure that the 2 key parts of an AES-XTS key are not the same
For security reasons, the 2 key parts of an AES-XTS key should not be the
same. Add checks to key generation and import to ensure this.

In case the random key generation results in 2 identical key blobs (very
unlikely), then the generation of the second key parts is repeated.
We compare the checksum, not the key blobs, because because the blobs are
encrypted and use a different IV, so it can't be determined from the blob
as such if the keys are identical. However, the checksum would be identical
if 2 identical keys would have been generated.

In case an AES-XTS key in imported from clear via C_CreateObject, and the
2 key parts supplied via CKA_VALUE are identical, the operation fails with
CKR_ATTRIBUTE_VALUE_INVALID.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>