CCA: Ensure that the 2 key parts of an AES-XTS key are not the same
For security reasons, the 2 key parts of an AES-XTS key should not be the
same. Add checks to key generation and import to ensure this.

In case the random key generation results in 2 identical encrypted keys
(very unlikely), then the generation of the key parts is repeated.

In case an AES-XTS key in imported from clear via C_CreateObject, and the
2 key parts supplied via CKA_VALUE are identical, the operation fails with
CKR_ATTRIBUTE_VALUE_INVALID.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>