EP11: Also check WKVP of MACed SPKIs
MACed SPKIs also contain the wrapping key verification pattern (WKVP) of
the wrapping key (EP11 master key) they were created with.

When generating key pairs by means of m_GenerateKeyPair(), then it is
guaranteed that the resulting private key blob and the MACed SPKI use
the same wrapping key, and thus have the same WKVP. So it used to be
sufficient to check just the WKVP of the private key blob.

Now that the MACed SPKI that is a result of a m_GenerateKeyPair() may be
re-created (because of different EP11 sessions used), the recreation may run
on a different EP11 adapter, which (erroneously) may have a different
wrapping key. So the resulting MACed SPKI must now also be checked to have
the expected WKVP.

Furthermore, let's now also check the WKVP of MACed SPKIs created by
importing or deriving public keys.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>