p11sak: Add --so option to login as SO (security officer)
With the '--so' option, p11sak opens a R/W SO session.
As per PKCS#11 standard, there is no R/O SO session, so always open
a R/W session if SO logs in.

Option '-p'/'--pin' must specify the SO pin, or if the '-p'/'--pin' option
is not specified, environment variable PKCS11_SO_PIN is used.
If PKCS11_SO_PIN is not set, then the user is prompted for the SO PIN.

SO can only access public token objects (CKA_PRIVATE=FALSE) as per PKCS#11
standard.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>