p11sak: Fixes for the extract-cert-pubkey command
Correct the description of the --new-attr option to list all attributes that
are valid for a public key, not only those that are valid for a certificate.

Don't refer to the certificate type and label when parsing the attributes
for the to be created public key leads to no applicable attributes.

Don't increase num_failed twice when parsing the new attributes fails.
Also, when none of the specified attributes are applicable for public keys,
treat this as a failure, not as a skip. With this, the data parameter is
superfluous for function p11sak_extract_pubkey().

Minor man page corrections.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>