EP11: Also update IBM specific attributes for public key after key gen
After an EP11 key has been created, e.g. by means of key generation, derive,
unwrap, or import, a couple of IBM specific attributes are retrieved from the
key blob and are added to the key's template.

Do this also for the public key during key pair generation, it was erroneously
only done for the private key.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>