COMMON: Allow a token to check for object access on creation and use
A token can register a check_obj_access function that gets called whenever an
object gets created or used. It can then check if accessing this object is
allowed and if not, return an appropriate return code.

This can for example be used to restrict access to objects when a token
operates in FIPS mode and must allow access to public objects only when USER is
logged in.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>