EP11: Refactor VHSM mode handling
The VHSM Session is no longer bound to the PKCS#11 session, but is token
global. The VHSM session is independent of the PKCS#11 session, thus there
is no need to bind it to a PKCS#11 session. The VHSM session is the same
for all PKCS#11 sessions across all OpenCryptoki processes anyway.

The VHSM session is still established on PKCS#11 session login, but only
once for the very first PKCS#11 session not for every session. Similarly,
on PKCS#11 session logout, the VHSM session is only logged out when the
very last PKCS#11 session (of all currently running OpenCryptoki processes)
is logged out.

Logging out the VHSM session only after the very last PKCS#11 session is
logged out prevents the code from situations where a VHSM-session-bound
key used on another PKCS#11 session would cause it to re-login the VHSM
session. If the VHSM session is only logged out when the very last PKCS#11
session is logged out, then no such re-login situation can occur.

For strict-session mode sessions, the information required to logout the session
is stored in a HW feature object, so that such sessions can manually be logged out
using the pkcsep11_session tool in case the application crashes before it is able
to logout the sessions.

For the VHSM sessions no special care must be taken for application crashes. Even
if an application crashes without logging out the VHSM session, then the next
application that logs in a USER PKCS#11 session will also logout the VHSM session
when it logs out the PKCS#11 session, and no other application still has any USER
sessions logged in. Since the VHSM sessions are the same across all applications,
this implicitly also logs out the left over VHSM session from the crashed
application. So just invoking the pkcsep11_session tool with command 'show' or
'logout' will implicitly perform this cleanup, because it logs in and out a USER
session.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>