EP11: Lazy creation of wrap-blob for importing clear keys
For importing clear keys, the key material must first be encrypted by
an AES key, and then unwrapped to get the secure key. For this an
AES secure key (the so called wrap-blob) must be generated.

Formerly this wrap-blob was generated at token initialization. Create the
wrap-blob on first usage only in a thread save way. This allows that the
wrap-blob is session-bound (required with with FIPS-session mode), because
the pin-blob needed to make the wrap-blob session-bound is only available
during key import (when a session is logged in), but not at token
initialization.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>