EP11: Introduce FIPS_SESSION_MODE config option
Support keyword FIPS_SESSION_MODE in the EP11 token configuration file to
enable the FIPS-session mode. FIPS-session mode can not be enabled
together with the VHSM mode or strict-session mode.

In FIPS-session mode a USER session is required to be able to access any kind
of key objects, regardless of the setting of the CKA_PRIVATE and CKA_TOKEN
attributes. That is, a C_Login must be performed with the USER pin to be able
to create or use key objects. Furthermore all keys generated by the
EP11 token will strictly belong to the EP11 token that created it.
Every EP11 token requires a FIPS-pin to be set using the pkcsep11_session
tool. In FIPS-sesion mode, a FIPS approved algorithm is used to login the
session into the EP11 adapter(s). This requires certain firmware support
from the adapter(s) as well as from the EP11 host library.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>