EP11: Require USER session to access key objects in FIPS-session mode
FIPS-Session mode requires that all key objects are bound to an EP11 session.
An EP11 session can only be logged in when the FIPS pin is available. The FIPS
pin is stored in a private token HW object, and thus it can only be read when
USER is logged in. Because it is a private object, the payload is encrypted
when stored in the token directory.

Thus, in order to create or use key objects, a USER session is required.
This is a non-PKCS#11 standard restriction. As per PKCS#11, a public or a
SO session would be allowed to create and use public (CKA_PRIVATE=FALSE) key
objects. In FIPS-session mode this is no longer the case.

Other classes of objects (e.g. certificates) are not affected by this
restriction, and behave like the PKCS#11 standard states.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>