EP11: Check if APQN for FIPS session related control points
In case FIPS-session mode is NOT enabled, control points
XCP_CPB_ALLOW_NONSESSION (48) and XCP_CPB_ALLOW_LOGIN_PRE_F2021 (73)
can not be OFF, because not all keys will be session bound, and the
pre-FIPS-2021 login algorithm is used.

Fail token initialization if one or more APQNs has one or both of these
control points OFF and FIPS-session mode is not active.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>