EP11: Implement login and logoff of FIPS sessions
When the FIPS-session mode is enabled, login the FIPS session with the EP11
APQNs configured for the token whenever a USER logs in into a PKCS#11 session,
and logout the FIPS session with the EP11 APQNs when the PKCS#11 session is
logged out.

For the FIPS session no special care must be taken for application crashes.
Same as for the VHSM session, if an application crashes without logging out the
FIPS session, then the next application that logs in a USER PKCS#11 session
will also logout the FIPS session when it logs out the PKCS#11 session, and no
other application still has any USER sessions logged in. Since the FIPS session
is the same across all applications, this implicitly also logs out the left
over FIPS session from the crashed application. So just invoking the
pkcsep11_session tool with command 'show' or 'logout' will implicitly perform
this cleanup, because it logs in and out a USER session.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>