EP11: Lazy obtaining of the firmware master key verification pattern
To obtain the firmware master key verification pattern for PKEY support,
a temporary AES key needs to be generated and converted into a protected
key, so that the firmware master key verification pattern can be read
from it.

Formerly this AES-key was generated at token initialization. In FIPS-session
mode, all secure keys must be bound to a session. However, at token
initialization no session is available. Obtain the firmware master key
verification pattern only at the first time it is really needed. This then
happens when a session is available, and thus this temporary AES key can also
be session bound.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>