EP11: Implement the LoginExtended protocol
The EP11 m_LoginExtended() and m_LogOutExtended() functions require the
caller to provide certain information BER encoded. Also, the caller must
perform an EC-DH key exchange with the EP11 APQN to derive a shared
secret that is then used to encrypt the pin, as well as the resulting pin
blob. Furthermore, some post-processing must be performed on the pin blob
retrieved, to make an operational pin blob out of it.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>