EP11: Store list of adapter serial numbers in EP11 session HW feature object
When logging in a strict-session EP11 session, the information required to log
this session out is stored in a HW feature token object, so that the session
can be manually logged out using the pkcsep11_session tool in case the
application crashes before it can logout the EP11 session.

Add the list of adapter serial numbers where the EP11 session was logged in
to the CKA_APPLICATION attribute of the HW feature object. The object is
removed by the token's C_Logout handling, or the pkcsep11_session tool only,
if the session was logged out on all adapters it was logged in, based on the
list of serial numbers in the HW feature token object. That way, if an adapter
went offline or is no longer available for whatever reason, before the session
was logged out on it, the information is kept, and the session can still be
logged out at a later time when the adapter is available again.

The CKA_APPLICATION attribute of the EP11 session HW feature token object now
contains the APQNs configured for the token (fixed size), followed by a list
of 16 byte serial number entries. The size of the CKA_APPLICATION attribute
can be used to determine the number of serial number entries following.

The pkcsep11_session tool now also displays the serial numbers of the adapters
the session was logged in. For the 'logout' command, it attempts to logout the
session on all APQNs configured for the token, and removes the adapters serial
number from the list of serial numbers. Only if that list is empty after
processing all configured and online APQNs, the HW feature token object is
removed. If it was not able to logout the session on all adapters, i.e. there
are still one or multiple remaining serial numbers in the list, the HW feature
object is kept and updated with the remaining list of serial numbers, and the
session can be logged out at a later time. You can force to remove the HW
feature object using the '-force' option.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>