EP11: Set CKA_ALWAYS_SENSITIVE and CKA_NEVER_EXTRACTABLE correctly at keygen
When a secret key (CKO_SECRET_KEY) or a private key (CKO_PRIVATE_KEY) is
generated with CKA_SENSITIVE=TRUE and/or CKA_EXTRACTABLE=FALSE then set
CKA_ALWAYS_SENSITIVE to TRUE and/or CKA_NEVER_EXTRACTABLE to TRUE
accordingly.

This is the same as done by common code in key_mgr_generate_key() and
key_mgr_generate_key_pair(). Factor-out this code and make a separate
function out of it that can be called from the EP11 token.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>