EP11: Set CKA_SENSITIVE defaults dependent on FORCE_SENSITIVE option
The PKCS#11 standard defines that the default value for CKA_SENSITIVE is
FALSE for secret keys (CKO_SECRET_KEY). For private keys (CKO_PRIVATE_KEY)
the default is token dependent.

If the FORCE_SENSITIVE token option is enabled, set attribute CKA_SENSITIVE
to TRUE for secret keys (CKO_SECRET_KEY) if it is not specified in the
template. This is against the PKCS#11 standard, but it is done only when
the FORCE_SENSITIVE option is enabled.

For private keys (CKO_PRIVATE_KEY) always set CKA_SENSITIVE to TRUE if it
is not specified in the template, regardless of the FORCE_SENSITIVE option.
This is according to the PKCS#11 standard, since for private keys the
default is defined as being token dependent, and thus the EP11 token can
define its own default value (i.e. TRUE).

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>