TESTCASES: ec_tests: Don't try to compare derived keys for secure key tokens
Trying to get the clear key value of a key from a secure key token will not
reveal the clear key value, but just binary zeros. This is the case, even if
the key is not sensitive (CKA_SENSITIVE=FALSE). A secure key will never ever
reveal its clear key value.

Comparing the returned all zero key value of the 2 derived keys does not make
much sense. Compare the MAC (using HMAC-SHA1) using the keys instead.
If the keys are the same, the MAC must also be the same.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>