CCA: Add support for ECDH key derivation
Add support for ECDH key derivation for the CCA token using the
CKM_ECDH1_DERIVE mechanism.

CCA supports the CKM_ECDH1_DERIVE-style key derivation with CCA verb
CSNDEDH and keyword 'DERIV02'. However, the use of 'DERIV02' imposes
certain restrictions:

- Only version 1 ECC private key tokens can be used (keyword 'ECC-VER1'
  required during key generation).
- The key-usage field must permit key establishment (keywords 'KEY-MGMT' or
  'KM-ONLY' required during key generation).
- The ECC private key must have been generated by random (see pedigree/
  key-source field). Keys imported from clear are not allowed to be used
  for ECDH.
- The ECC private key token must include an ECC key derivation info section
  (X'23') that specifies the key type and size of the to-be-derived keys.
  Only exactly one key type and size can be specified there. The key
  derivation info must be supplied during key generation.

This means that ECDH key derivation using existing ECC private keys, that
have been generated with an earlier version of the CCA token will not allow
to perform ECDH (even though they might have CKA_DERIVE=TRUE). Doing so will
fail with CKR_KEY_FUNCTION_NOT_PERMITTED.

With this commit, ECC Keys generated with the CCA token will fulfill the
requirements to be able to perform ECDH key derivation.

The type and size of the to-be-derived keys must be specified at key generation
time already. The CCA token inspects the CKA_DERIVE_TEMPLATE specified for
the private key and extracts the CKA_KEY_TYPE, CKA_VALUE_LEN and
CKA_IBM_CCA_AES_KEY_MODE attributes from it (if specified). If the key type
and the value length are contained, then the CCA token will generate an ECC key
with the ECC key derivation info section containing the specified key type
and size. The only supported key type is AES. By default, the CCA token will
generate ECC keys that allow to derive AES-256 keys. If the key mode is
contained in the derive template, then the ECC key derivation info section
will contain the desired key mode. If the key mode is not contained, then
the token global default mode is used.

Keys that are imported from clear (via C_CreateObject) will not be able to
support ECDH, because they can not fulfill the requirements (i.e. not
generated by random). Consequently, the CCA token does not allow CKA_DERIVE
to be TRUE for ECC keys imported from clear. Doing so will fail with
CKR_TEMPLATE_INCONSISTENT.

Existing CCA key blobs can be imported into the CCA token by using
C_CreateObject and supplying the key blob with the CKA_IBM_OPAQUE attribute.
If CKA_DERIVE is TRUE for such a key, it is checked if the key blob fulfills
the requirements for ECDH. Furthermore, it is checked that the to-be-derived
key mode is DATA or CIPHER. In case the requirements are not fulfilled, the
import fails with CKR_ATTRIBUTE_VALUE_INVALID.

Trying to derive key types and sizes other than those allowed per ECC key
derivation info section of the private key used will cause the C_DeriveKey
call to fail with CKR_KEY_TYPE_INCONSISTENT or CKR_ATTRIBUTE_VALUE_INVALID
(key size invalid for AES keys).

If the attribute CKA_DERIVE is being set to TRUE on an existing ECC private
key object, is checked that the CCA ECC private key token contains a
ECC key derivation info section with valid information. Only then setting
CKA_DERIVE to TRUE is allowed, because only then an ECDH key derivation can
be performed with this key. Otherwise CKR_ATTRIBUTE_VALUE_INVALID is
returned.

If the attribute CKA_DERIVE_TEMPLATE is being changed on an existing ECC
private key object, it is checked that if the CCA ECC private key token
contains a ECC key derivation info section it matches to the key type,
key size, and key mode specified in the derive template. Otherwise
CKR_ATTRIBUTE_VALUE_INVALID is returned.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>