TESTCASE: Correct test run counting for some test cases
When running the digest tests against the EP11 token, the number of passed
tests is greater than the number of ran tests:

  Total=849, Ran=365, Passed=389, Failed=0, Skipped=484, Errors=0

The same happens for the dilithium and kyber test cases, when some dilithium
or kyber variants are not supported by the crypto cards.

Fix this by adding a call to testcase_new_assertion() at the right place
to ensure the ran counter is increased before a test is counted as being
passed or failed.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>