TESTCASES: ec_tests: Support ECDH KATs for secure key tokens
For secure key tokens, the derived secret can not be compared with the expected
key value for known answer tests (KATs), because the derived secret is a secure
key that never reveals its key value in clear.

However, the test can calculate an HMAC of some known data using the derived key
as well as with the expected key value from the KAT. If both keys are the same,
then the calculated HMAC must also be the same.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>