EP11: Support ECDH with compressed or hybrid EC point as public data
The EP11 host library does not seem to like compressed or hybrid EC points
as public data with CKM_ECDH1_DERIVE (fields pPublicData/ulPublicDataLen in
CK_ECDH1_DERIVE_PARAMS).

Uncompress it in case it is in compressed form and convert it to
uncompressed form in case it is in hybrid form.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>