ICSF: Cache attributes for policy attribute retrieval
Policy checks on objects require to retrieve certain attributes of the
object, dependent on the object type.

Retrieving attributes is a relatively time consuming operation for the ICSF
token, since it needs to fetch all attributes of the object from ICSF and
then return only that one attribute that was asked for.

When multiple attributes are to be retrieved one after the other, then the
same processing takes place again and again.

Cache the retrieved attributes of an object on the first policy attribute
retrieval, so that subsequent attribute retrievals can be answered quickly
from the cache.

The cache is cleared and freed when the very last policy attribute is freed.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>