ICSF: Fix attribute retrieval handling
Function C_GetAttributeValue is supposed to set ulValueLen to
CK_UNAVAILABLE_INFORMATION if an attribute was not found, or the buffer
supplied with pValue is too small. If the pValue of an attribute is NULL,
then ulValueLen is set to the required buffer size of the attribute value.
If pValue is not NULL, then ulValueLen specifies the size of the supplied
buffer. If the size of the attribute value is larger than the supplied buffer
size, then the attribute value is not copied to the buffer, ulValueLen is set
to CK_UNAVAILABLE_INFORMATION and the function returns CKR_BUFFER_TOO_SMALL.

The code did preset ulValueLen of all attributes to CK_UNAVAILABLE_INFORMATION
priory to decoding the retrieved attributes. This overwrites any attribute
value buffer size specified by the caller. So the code can't check if the
supplied buffer is large enough to hold the value of the attribute.

Don't preset ulValueLen with CK_UNAVAILABLE_INFORMATION and thus retain the
buffer size specified by the caller. This allows a proper length check when
decoding the attributes. In case not all attributes have been found, then
set ulValueLen to CK_UNAVAILABLE_INFORMATION for those attributes that have
not been found, and thus have not been updated.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>