policy: Fix a use-after-free during retrieval of EC curve information
Don't pass back the pValue pointer from the retrieved attribute. The getattr
callback might have allocated it and thus the free_attr callback frees it
before returning from this function. This then returns a pointer to an just
freed memory block.

The code anyway looks up the curve in the static const curve info array
(der_ec_supported), so return the pointer to the OID inside that array. This
is a pointer to a static const memory block, and thus can be returned safely.

Currently only the ICSF token uses a getattr callback that allocates the
attribute. The other tokens just return the attribute that is linked from
within the object's template. So only the ICSF token causes a use-after-free
error.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>