pin_prompt: Fix heap-buffer-overflow if pin_prompt() returns NULL
When getline() returns -1 to indicate an error, it may anyway have allocated
the buffer. When that buffer is later freed using pin_free(), it tries to
clean the buffer and uses strlen() to get the amount of characters to clean.
However, if getline() failed, it just allocated the buffer, but did not
initialize it, so the use of strlen() may overrun the buffer and thus
pin_free() attempts to clean bytes beyond the allocated buffer.

Fix this by freeing the buffer and setting the pointer to NULL in case
getline() failed. This also removes the necessity that the caller frees
the buffer via pin_free() even if pin_prompt() returned NULL.

Suggested-by: Holger Dengler <dengler@linux.ibm.com>
Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>